requirements.txt
absl-py==1.4.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.2
bitarray==2.7.4
black==23.3.0
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.0
ccimport==0.4.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
colorama==0.4.6
colour-demosaicing==0.2.3
colour-science==0.4.2
comm==0.1.3
ConfigArgParse==1.5.3
contourpy==1.0.7
cumm-cu120==0.4.9
cycler==0.11.0
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
-e git+https://github.com/facebookresearch/detectron2.git@2c6c380f94a27bd8455a39506c9105f652b9f760#egg=detectron2
exceptiongroup==1.1.1
execnet==1.9.0
executing==1.2.0
faiss-cpu==1.7.4
fastjsonschema==2.16.3
filelock==3.12.0
fire==0.5.0
flake8==6.0.0
Flask==2.3.1
fonttools==4.39.3
fqdn==1.5.1
fvcore==0.1.5.post20221221
google-auth==2.18.1
google-auth-oauthlib==1.0.0
grpcio==1.55.0
h5py==3.8.0
hydra-core==1.3.2
idna==3.4
imageio==2.28.0
iniconfig==2.0.0
install==1.3.5
iopath==0.1.9
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lark==1.1.5
lazy_loader==0.2
lit==16.0.3
llvmlite==0.40.0
lyft-dataset-sdk==0.0.8
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
mmcv==2.0.0
mmdet==3.0.0
mmdet3d==1.1.0
mmengine==0.7.3
model-index==0.1.11
mpmath==1.3.0
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.0
numpy==1.24.3
nuscenes-devkit==1.1.10
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.17.0
opencv-python==4.7.0.72
openmim==0.3.7
ordered-set==4.1.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pccm==0.4.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.4.0
plotly==5.14.1
pluggy==1.0.0
plyfile==0.9
portalocker==2.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
pytest==7.3.1
pytest-xdist==3.2.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-metric-learning==0.9.94
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
requests==2.30.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.2
Shapely==1.8.5
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
spconv-cu120==2.3.6
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
terminado==0.17.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1
torch-tb-profiler==0.4.1
torchinfo==1.8.0
torchvision==0.15.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
trimesh==3.21.7
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==2.3.0
widgetsnbextension==4.0.7
yacs==0.1.8
yapf==0.33.0
